backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
pydantic==2.6.0
pandas==2.2.2
python-multipart==0.0.6
aiohttp==3.9.4
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.44
databases==0.9.0